requirements.txt
torch>=2.0.0
numpy>=1.24
scipy>=1.10
soundfile>=0.10
PyYAML>=6.0
tqdm>=4.60
